Refactor: using dedicated dir for the whole functionality
