+++ webapp/src/__tests__/App.test.js
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import App from '../App';
 import { SessionContext } from '../SessionContext';
-import '../i18n';
+import '../localize/i18n';
 
 describe('App component', () => {
   it('renders home page by default', () => {
