+++ webapp/src/__tests__/components/NavBar.test.js
@@ -3,7 +3,7 @@ import { render, screen, waitFor, fireEvent } from '@testing-library/react';
 import { BrowserRouter  } from 'react-router-dom';
 import NavBar from '../../components/NavBar';
 import { SessionContext } from '../../SessionContext';
-import '../../i18n';
+import '../../localize/i18n';
 
 describe('NavBar component', () => {
   it('should render logo when not logged', async () => {
